app/requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
Pillow==9.5.0